Added server-side python test script to receive audio
To use the script, upload it to a server and run it using python.  It
will listed on port 4444 for incoming POSTs containing base64-encoded
bodies.  It will decode the body and save it to a file.

Example of sending data to it from Electric Imp agent code:

local req = http.post("http://bobert.net:4444",
                      {"Content-Type": "application/octet-stream"},
                      base64_encode(gAudioBuffer));