Automate STM32 software version check, upgrade when going to sleep
The STM32 software version is recorded on every speak/scan when the
STM32 is started and sends a packet with its software version. If the
software version does not match, the STM32 software is upgraded to the
software embedded in the imp device model just prior to the device going
to sleep.