WIP: Instant on
this now saves the audio data into flash when there is no connection
and as soon as the connection is available it would send it to the
server.

if the button is released after recording and still no connection, the
code will wait until the connection is available and would send the
pending audio buffer to agent and then onto server.