improving the factory test to reliably pass audio test
this one seems to work just fine.  but I would like to understand why
it fails when the test is ran with all four of them it fails.

I have a theory where if you place the device right across from the
barcode in the compartment then it fires the scanner because it
recognized a barcode.